+++ packages/components/CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### [Unreleased]
 
+### [2.9.2] - 2022-03-23
+
 ### [2.9.2-rc.1] - 2022-03-23
 
 #### Fixed